modularised contexts and providers
